Split manifest generator into composed subgenerators.
